+++ src/components/User/ShowProfileLinks/ShowProfileLinks.tsx
@@ -51,5 +51,48 @@ export const CustomLink = ({ customLinks, index }: customLinkProps) => {
 };
 
 export const AddProfileLinks = () => {
-  return <div>ShowProfileLinks</div>;
+  const [customInputs, setCustomInputs] = useState<any>([]);
+
+  const addCustomInput = () => {
+    setCustomInputs((link: any) => {
+      return [
+        ...link,
+        {
+          type: 'text',
+          value: '',
+        },
+      ];
+    });
+  };
+
+  const handleCustomInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    const { name, value, id } = e.target;
+    console.log(id);
+    setCustomInputs((link: any) => {
+      const newCustomInputArr = link.slice();
+      newCustomInputArr[id].value = value;
+      console.log(newCustomInputArr[id].value);
+      return newCustomInputArr;
+    });
+  };
+  return {
+    /* <button type="button" onClick={addCustomInput}>
+  + Add Custom Link
+</button>
+
+{customInputs.map((link: any, index: any) => {
+  return (
+    <div key={index}>
+      <label htmlFor="customUrlName">
+        Url Name
+        <input id={index} type={link.type} name="customUrlName" onChange={handleCustomInputChange} required />
+      </label>
+      <label htmlFor="customUrlLink">
+        Url Link
+        <input id={index} type={link.type} name="customUrlLink" onChange={handleCustomInputChange} required />
+      </label>
+    </div>
+  );
+})} */
+  };
 };
